% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{Print a friendly greeting}
\usage{
hello_world()
}
\value{
Invisibly returns the printed string \code{"Hello, world!"}.
}
\description{
Print a friendly greeting
}
\examples{
hello_world()
}
