% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world2}
\alias{hello_world2}
\title{Print a second friendly greeting}
\usage{
hello_world2()
}
\value{
Invisibly returns the printed string \code{"Hello, world2!"}.
}
\description{
Print a second friendly greeting
}
\examples{
hello_world2()
}
